execdetails: Move logic from execdetails.go (2468 lines) to runtime_stats.go, tiflash_stats.go, and util.go
